fixed margins and centered icons
